requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Django==4.1.7
et-xmlfile==1.1.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
multidict==6.0.4
networkx==3.0
numpy==1.24.2
pandas==1.5.3
pydantic==1.10.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
typing_extensions==4.5.0
uvicorn==0.20.0
yarl==1.8.2
